requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pydantic==1.10.17
google-api-python-client==2.141.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1